refactor: remove dead tab views, add balance activity heatmap (AND-312, WIP)
