--- Sources/PlaidBar/Views/RecurringView.swift
@@ -1,122 +1,0 @@
-import SwiftUI
-import PlaidBarCore
-
-struct RecurringView: View {
-    @Environment(AppState.self) private var appState
-
-    private var emptyPresentation: SecondaryContentUnavailableState {
-        SecondaryContentUnavailableState.recurring(
-            isDemoMode: appState.isDemoMode,
-            isInitialLoad: appState.loadState(for: .recurring).isInitialLoad,
-            serverConnected: appState.serverConnected,
-            linkedItemCount: appState.statusItemCount,
-            accountCount: appState.accounts.count,
-            syncedItemCount: appState.serverSyncedItemCount ?? 0,
-            transactionCount: appState.transactions.count,
-            errorMessage: appState.error
-        )
-    }
-
-    var body: some View {
-        let recurring = appState.recurringTransactions
-        let monthlyEstimate = appState.estimatedMonthlyRecurring
-
-        VStack(alignment: .leading, spacing: 0) {
-            if recurring.isEmpty {
-                emptyState
-            } else {
-                // Estimated monthly total header (normalizes weekly/annual to monthly equivalent)
-                HStack {
-                    Text("EST. MONTHLY COST")
-                        .sectionTitle()
-                        .foregroundStyle(.secondary)
-                    Spacer()
-                    Text(Formatters.currency(monthlyEstimate, format: .full))
-                        .heroBalance()
-                }
-                .padding(.horizontal, Spacing.lg)
-                .padding(.top, Spacing.md)
-                .padding(.bottom, Spacing.sm)
-                .accessibilityElement(children: .combine)
-                .accessibilityLabel("Estimated monthly recurring cost: \(Formatters.currency(monthlyEstimate, format: .full))")
-
-                Divider()
-
-                ForEach(recurring) { item in
-                    RecurringRow(item: item)
-                }
-            }
-        }
-    }
-
-    @ViewBuilder
-    private var emptyState: some View {
-        SecondaryUnavailableView(presentation: emptyPresentation) {
-            performEmptyAction(emptyPresentation.action)
-        }
-    }
-
-    private func performEmptyAction(_ action: SecondaryContentUnavailableAction) {
-        switch action {
-        case .checkServer:
-            Task { await appState.checkServerConnection() }
-        case .addAccount:
-            Task { await appState.addAccount() }
-        case .refreshAccounts:
-            Task { await appState.refreshAccounts() }
-        case .syncTransactions:
-            Task { await appState.syncTransactions() }
-        case .refresh:
-            Task { await appState.refreshDashboard() }
-        case .clearFilters, .showWiderPeriod:
-            break
-        }
-    }
-}
-
-private struct RecurringRow: View {
-    let item: RecurringTransaction
-
-    var body: some View {
-        HStack(spacing: Spacing.md) {
-            // Category icon
-            Image(systemName: (item.category ?? .other).iconName)
-                .font(.body)
-                .foregroundStyle(.secondary)
-                .frame(width: 24)
-
-            VStack(alignment: .leading, spacing: Spacing.xxs) {
-                Text(item.merchantName)
-                    .font(.body)
-                    .lineLimit(1)
-
-                HStack(spacing: Spacing.sm) {
-                    // Frequency badge
-                    Text(item.frequency.displayName)
-                        .microText()
-                        .padding(.horizontal, Spacing.sm)
-                        .padding(.vertical, Spacing.xxs)
-                        .background(SemanticColors.recurring.opacity(0.15), in: Capsule())
-                        .foregroundStyle(SemanticColors.recurring)
-
-                    Text(Formatters.currency(item.averageAmount, format: .full))
-                        .detailText()
-                        .monospacedDigit()
-                }
-
-                Text("Last: \(Formatters.displayTransactionDate(item.lastDate)) • Next: \(Formatters.displayTransactionDate(item.nextExpectedDate))")
-                    .detailText()
-
-                Text("\(item.transactionCount) matching charges • \(Formatters.percent(item.confidence * 100, decimals: 0)) confidence")
-                    .detailText()
-            }
-
-            Spacer()
-        }
-        .padding(.horizontal, Spacing.lg)
-        .padding(.vertical, Spacing.sm)
-        .hoverHighlight()
-        .accessibilityElement(children: .combine)
-        .accessibilityLabel("\(item.merchantName), \(item.frequency.displayName), \(Formatters.currency(item.averageAmount, format: .full)), \(item.transactionCount) matching charges, \(Formatters.percent(item.confidence * 100, decimals: 0)) confidence, next expected \(Formatters.displayTransactionDate(item.nextExpectedDate))")
-    }
-}
