--- Sources/PlaidBar/Views/StatusView.swift
@@ -1,441 +1,0 @@
-import AppKit
-import PlaidBarCore
-import SwiftUI
-
-struct StatusView: View {
-    @Environment(AppState.self) private var appState
-    @Environment(\.openSettings) private var openSettings
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: Spacing.lg) {
-            statusHeader
-
-            AttentionQueueView(title: "ATTENTION")
-                .environment(appState)
-
-            localDataDirectoryRow
-
-            diagnosticsGrid
-
-            if !appState.itemStatuses.isEmpty {
-                itemSection
-            }
-
-            recoveryActions
-        }
-        .padding(Spacing.lg)
-    }
-
-    private var statusHeader: some View {
-        HStack(spacing: Spacing.md) {
-            Image(systemName: statusIcon)
-                .font(.system(size: 28, weight: .semibold))
-                .foregroundStyle(statusColor)
-                .frame(width: 36)
-
-            VStack(alignment: .leading, spacing: Spacing.xs) {
-                Text(appState.diagnosticsSummary)
-                    .font(.headline)
-                    .lineLimit(1)
-                    .minimumScaleFactor(0.8)
-
-                Text("Mode: \(appState.statusModeText)")
-                    .detailText()
-            }
-
-            Spacer()
-        }
-        .padding(Spacing.md)
-        .background(.quaternary.opacity(0.5))
-        .clipShape(RoundedRectangle(cornerRadius: 8))
-    }
-
-    private var localDataDirectoryRow: some View {
-        HStack(alignment: .top, spacing: Spacing.sm) {
-            Image(systemName: "internaldrive")
-                .font(.callout.weight(.semibold))
-                .foregroundStyle(.secondary)
-                .frame(width: 26, height: 26)
-                .background(.quaternary.opacity(0.45), in: RoundedRectangle(cornerRadius: 7))
-                .accessibilityHidden(true)
-
-            VStack(alignment: .leading, spacing: Spacing.xxs) {
-                Text("Active data directory")
-                    .font(.caption.weight(.semibold))
-
-                Text(appState.activeStorageDirectoryDisplayText)
-                    .microText()
-                    .foregroundStyle(.secondary)
-                    .lineLimit(1)
-                    .truncationMode(.middle)
-
-                Text("Local-only diagnostics; credentials and Plaid IDs are not shown.")
-                    .microText()
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-
-            Spacer(minLength: Spacing.xs)
-        }
-        .padding(.horizontal, Spacing.sm)
-        .padding(.vertical, Spacing.rowVertical)
-        .nativePanelSurface()
-        .accessibilityElement(children: .combine)
-        .accessibilityLabel(
-            "Active data directory: \(appState.activeStorageDirectoryDisplayText). Local-only diagnostics; credentials and Plaid IDs are not shown."
-        )
-    }
-
-    private var diagnosticsGrid: some View {
-        LazyVGrid(
-            columns: [
-                GridItem(.flexible(), spacing: Spacing.sm),
-                GridItem(.flexible(), spacing: Spacing.sm),
-            ],
-            spacing: Spacing.sm
-        ) {
-            DiagnosticTile(
-                title: "Server",
-                value: appState.statusServerText,
-                icon: serverDiagnosticIcon,
-                color: serverDiagnosticColor
-            )
-            DiagnosticTile(
-                title: "Sync",
-                value: appState.statusSyncText,
-                icon: "arrow.triangle.2.circlepath",
-                color: appState.isSyncStale ? SemanticColors.warning : SemanticColors.positive
-            )
-            DiagnosticTile(
-                title: "Accounts",
-                value: "\(appState.accountCount)",
-                icon: "creditcard",
-                color: SemanticColors.brand
-            )
-            DiagnosticTile(
-                title: "Transactions",
-                value: "\(appState.transactionCount)",
-                icon: "list.bullet.rectangle",
-                color: SemanticColors.brandSecondary
-            )
-            DiagnosticTile(
-                title: "Plaid items",
-                value: "\(appState.connectedItemCount)/\(appState.statusItemCount)",
-                icon: "link",
-                color: itemHealthColor
-            )
-            DiagnosticTile(
-                title: "Credentials",
-                value: appState.serverCredentialsText,
-                icon: "key",
-                color: credentialsColor
-            )
-            DiagnosticTile(
-                title: "Sync ready",
-                value: appState.serverSyncReadinessText,
-                icon: "checklist.checked",
-                color: syncReadinessColor
-            )
-            DiagnosticTile(
-                title: "Version",
-                value: appState.serverVersion ?? PlaidBarConstants.appVersion,
-                icon: "number",
-                color: .secondary
-            )
-            DiagnosticTile(
-                title: "Endpoint",
-                value: appState.localServerURLText.replacingOccurrences(of: "http://", with: ""),
-                icon: "network",
-                color: .secondary
-            )
-            DiagnosticTile(
-                title: "Refresh",
-                value: appState.refreshCadenceText,
-                icon: "timer",
-                color: .secondary
-            )
-        }
-    }
-
-    private var itemSection: some View {
-        VStack(alignment: .leading, spacing: Spacing.sm) {
-            Text("PLAID ITEMS")
-                .sectionTitle()
-                .foregroundStyle(.secondary)
-
-            VStack(spacing: Spacing.xs) {
-                ForEach(appState.itemStatuses) { item in
-                    HStack(spacing: Spacing.sm) {
-                        Image(systemName: icon(for: item.status))
-                            .font(.callout.weight(.semibold))
-                            .foregroundStyle(color(for: item.status))
-                            .frame(width: 26, height: 26)
-                            .background(color(for: item.status).opacity(0.12), in: RoundedRectangle(cornerRadius: 7))
-
-                        VStack(alignment: .leading, spacing: Spacing.xxs) {
-                            Text(item.institutionName ?? "Plaid item")
-                                .font(.callout.weight(.medium))
-                                .lineLimit(1)
-
-                            Text(statusDetail(for: item))
-                                .detailText()
-                        }
-
-                        Spacer()
-
-                        VStack(alignment: .trailing, spacing: Spacing.xs) {
-                            Text(label(for: item.status))
-                                .microText()
-                                .padding(.horizontal, 7)
-                                .padding(.vertical, 4)
-                                .foregroundStyle(color(for: item.status))
-                                .background(color(for: item.status).opacity(0.12), in: Capsule())
-
-                            if item.status != .connected {
-                                Button {
-                                    Task { await appState.reconnectItem(itemId: item.id) }
-                                } label: {
-                                    Label("Reconnect", systemImage: "link.badge.plus")
-                                }
-                                .buttonStyle(.bordered)
-                                .controlSize(.mini)
-                            }
-                        }
-                    }
-                    .padding(Spacing.sm)
-                    .background(itemBackground(for: item.status), in: RoundedRectangle(cornerRadius: 8))
-                    .accessibilityElement(children: .combine)
-                    .accessibilityLabel(itemAccessibilityLabel(for: item))
-                }
-            }
-        }
-    }
-
-    private var recoveryActions: some View {
-        let readiness = appState.dashboardStatusReadiness
-
-        return VStack(alignment: .leading, spacing: Spacing.sm) {
-            Text("RECOVERY")
-                .sectionTitle()
-                .foregroundStyle(.secondary)
-
-            if let primaryAction = readiness.primaryAction {
-                if readiness.level == .healthy {
-                    Button {
-                        perform(readinessAction: primaryAction)
-                    } label: {
-                        Label(
-                            primaryActionLabel(for: primaryAction, readiness: readiness),
-                            systemImage: readiness.primaryActionIconName ?? primaryAction.defaultIconName
-                        )
-                    }
-                    .buttonStyle(.bordered)
-                    .disabled(appState.isLoading)
-                } else {
-                    Button {
-                        perform(readinessAction: primaryAction)
-                    } label: {
-                        Label(
-                            primaryActionLabel(for: primaryAction, readiness: readiness),
-                            systemImage: readiness.primaryActionIconName ?? primaryAction.defaultIconName
-                        )
-                    }
-                    .buttonStyle(.borderedProminent)
-                    .tint(readinessTint(for: readiness.level))
-                    .disabled(appState.isLoading)
-                }
-            }
-        }
-    }
-
-    private var statusIcon: String {
-        if appState.isDemoStatusRecoveryScenario, appState.erroredItemCount > 0 || appState.needsLoginItemCount > 0 {
-            return "exclamationmark.triangle.fill"
-        }
-        if appState.isDemoMode { return "play.circle.fill" }
-        if !appState.serverConnected { return "xmark.octagon.fill" }
-        if appState.erroredItemCount > 0 || appState.needsLoginItemCount > 0 { return "exclamationmark.triangle.fill" }
-        return "checkmark.seal.fill"
-    }
-
-    private var statusColor: Color {
-        if appState.isDemoStatusRecoveryScenario, appState.erroredItemCount > 0 || appState.needsLoginItemCount > 0 {
-            return SemanticColors.warning
-        }
-        if appState.isDemoMode { return SemanticColors.brandSecondary }
-        if !appState.serverConnected { return SemanticColors.negative }
-        if appState.erroredItemCount > 0 || appState.needsLoginItemCount > 0 { return SemanticColors.warning }
-        return SemanticColors.positive
-    }
-
-    private var serverDiagnosticIcon: String {
-        appState.isDemoMode ? "play.circle.fill" : "server.rack"
-    }
-
-    private var serverDiagnosticColor: Color {
-        if appState.isDemoMode { return SemanticColors.brandSecondary }
-        return appState.serverConnected ? SemanticColors.positive : SemanticColors.negative
-    }
-
-    private var itemHealthColor: Color {
-        if appState.statusItemCount == 0 { return .secondary }
-        if appState.connectedItemCount < appState.statusItemCount { return SemanticColors.warning }
-        if appState.erroredItemCount > 0 || appState.needsLoginItemCount > 0 { return SemanticColors.warning }
-        return SemanticColors.positive
-    }
-
-    private var credentialsColor: Color {
-        if appState.isDemoMode { return .secondary }
-        guard appState.serverConnected else { return .secondary }
-        return appState.serverCredentialsConfigured == true ? SemanticColors.positive : SemanticColors.negative
-    }
-
-    private var syncReadinessColor: Color {
-        if appState.isDemoMode { return SemanticColors.brandSecondary }
-        guard appState.serverConnected else { return .secondary }
-        return appState.serverSyncReady == true ? SemanticColors.positive : .secondary
-    }
-
-    private func icon(for status: ItemConnectionStatus) -> String {
-        switch status {
-        case .connected: "checkmark.circle.fill"
-        case .loginRequired: "person.crop.circle.badge.exclamationmark"
-        case .error: "xmark.octagon.fill"
-        }
-    }
-
-    private func color(for status: ItemConnectionStatus) -> Color {
-        switch status {
-        case .connected: SemanticColors.positive
-        case .loginRequired: SemanticColors.warning
-        case .error: SemanticColors.negative
-        }
-    }
-
-    private func itemBackground(for status: ItemConnectionStatus) -> Color {
-        switch status {
-        case .connected:
-            SemanticColors.positive.opacity(0.045)
-        case .loginRequired:
-            SemanticColors.warning.opacity(0.08)
-        case .error:
-            SemanticColors.negative.opacity(0.08)
-        }
-    }
-
-    private func label(for status: ItemConnectionStatus) -> String {
-        switch status {
-        case .connected: appState.isDemoStatusRecoveryScenario ? "Recovered" : "Synced"
-        case .loginRequired: "Needs login"
-        case .error: "Error"
-        }
-    }
-
-    private func statusDetail(for item: ItemStatus) -> String {
-        switch item.status {
-        case .connected:
-            if appState.isDemoStatusRecoveryScenario {
-                item.lastSync.map { "Recovered and synced \(Formatters.relativeDate($0))" } ?? "Connected; no sync recorded"
-            } else {
-                item.lastSync.map { "Updated \(Formatters.relativeDate($0))" } ?? "No sync recorded"
-            }
-        case .loginRequired:
-            "Plaid requires a fresh bank login. Reconnect this item."
-        case .error:
-            "The last Plaid request failed. Reconnect or try refreshing again."
-        }
-    }
-
-    private func itemAccessibilityLabel(for item: ItemStatus) -> String {
-        let institutionName = item.institutionName ?? "Plaid item"
-        return "\(institutionName), \(label(for: item.status)), \(statusDetail(for: item))"
-    }
-
-    private func perform(readinessAction action: DashboardStatusReadinessAction) {
-        switch action {
-        case .checkServer:
-            Task { await appState.checkServerConnection() }
-        case .addAccount:
-            Task { await appState.addAccount() }
-        case .refresh:
-            Task { await appState.refreshDashboard() }
-        case .reconnect:
-            guard let itemId = ItemRecoveryTarget.itemId(from: appState.itemStatuses) else {
-                Task { await appState.refreshDashboard() }
-                return
-            }
-            Task { await appState.reconnectItem(itemId: itemId) }
-        case .openSettings:
-            openSettings()
-        case .requestNotificationPermission:
-            Task { _ = await appState.requestNotificationPermission() }
-        case .openNotificationSettings:
-            openNotificationSettings()
-        }
-    }
-
-    private func openNotificationSettings() {
-        guard let url = URL(string: "x-apple.systempreferences:com.apple.Notifications-Settings.extension") else {
-            openSettings()
-            return
-        }
-        NSWorkspace.shared.open(url)
-    }
-
-    private func primaryActionLabel(
-        for action: DashboardStatusReadinessAction,
-        readiness: DashboardStatusReadiness
-    ) -> String {
-        if action == .reconnect,
-           let title = ItemRecoveryTarget.actionTitle(from: appState.itemStatuses)
-        {
-            return title
-        }
-        return readiness.primaryActionTitle ?? action.defaultTitle
-    }
-
-    private func readinessTint(for level: DashboardStatusReadinessLevel) -> Color {
-        switch level {
-        case .healthy:
-            .accentColor
-        case .loading:
-            .secondary
-        case .warning:
-            SemanticColors.warning
-        case .blocked:
-            SemanticColors.negative
-        }
-    }
-}
-
-private struct DiagnosticTile: View {
-    let title: String
-    let value: String
-    let icon: String
-    let color: Color
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: Spacing.sm) {
-            HStack {
-                Image(systemName: icon)
-                    .foregroundStyle(color)
-                Spacer()
-            }
-
-            Text(value)
-                .font(.callout.weight(.semibold))
-                .lineLimit(1)
-                .minimumScaleFactor(0.75)
-
-            Text(title)
-                .detailText()
-                .lineLimit(1)
-        }
-        .frame(maxWidth: .infinity, minHeight: 78, alignment: .topLeading)
-        .padding(Spacing.sm)
-        .background(.quaternary.opacity(0.35))
-        .clipShape(RoundedRectangle(cornerRadius: 8))
-        .accessibilityElement(children: .combine)
-        .accessibilityLabel("\(title): \(value)")
-    }
-}
